requirements.txt
# requirements.txt
flask==2.0.1
flask-jwt-extended==4.3.1
pymongo[srv]
python-dotenv==0.19.1
werkzeug==2.0.1